Update index.htm
Changed the way it picks a map, so that it removes the map it chose, and adds all maps to the list  again for next time it picks a map. This way you can't get the map you just had, and maps become more likely to be picked, when not picked, and less likely when picked.